+++ unstoppable-ios-app/domains-manager-ios/Modules/Explore/Subviews/Suggested/HomeExploreSuggestedProfileRowView.swift
@@ -7,22 +7,21 @@
 
 import SwiftUI
 
-struct HomeExploreSuggestedProfileRowView: View {
+struct HomeExploreSuggestedProfileRowView: View, ViewAnalyticsLogger {
     
+    @EnvironmentObject var viewModel: HomeExploreViewModel
     @Environment(\.imageLoadingService) private var imageLoadingService
+    @Environment(\.analyticsViewName) var analyticsName
     
     let profileSuggestion: DomainProfileSuggestion
     
     @State private var avatar: UIImage?
     
     var body: some View {
-        HStack(spacing: 16) {
-            avatarView()
-            HStack(spacing: 2) {
-                profileInfoView()
-                Spacer()
-                actionButtonView()
-            }
+        HStack(spacing: 2) {
+            selectableProfileInfoView()
+            Spacer()
+            actionButtonView()
         }
         .frame(maxWidth: .infinity)
         .frame(height: 44)
@@ -53,6 +52,25 @@ private extension HomeExploreSuggestedProfileRowView {
 // MARK: - Private methods
 private extension HomeExploreSuggestedProfileRowView {
     @ViewBuilder
+    func selectableProfileInfoView() -> some View {
+        Button {
+            UDVibration.buttonTap.vibrate()
+            viewModel.didSelectDomainProfileSuggestion(profileSuggestion)
+        } label: {
+            profileInfoView()
+        }
+        .buttonStyle(.plain)
+    }
+    
+    @ViewBuilder
+    func profileInfoView() -> some View {
+        HStack(spacing: 16) {
+            avatarView()
+            profileDetailsInfoView()
+        }
+    }
+    
+    @ViewBuilder
     func avatarView() -> some View {
         ZStack(alignment: .bottomTrailing) {
             UIImageBridgeView(image: avatar)
@@ -78,7 +96,7 @@ private extension HomeExploreSuggestedProfileRowView {
     }
     
     @ViewBuilder
-    func profileInfoView() -> some View {
+    func profileDetailsInfoView() -> some View {
         VStack(alignment: .leading) {
             Text(profileSuggestion.domain)
                 .font(.currentFont(size: 16, weight: .medium))
@@ -96,7 +114,7 @@ private extension HomeExploreSuggestedProfileRowView {
     func actionButtonView() -> some View {
         UDButtonView(text: String.Constants.follow.localized(),
                      style: .small(.raisedPrimary)) {
-            
+            viewModel.didSelectToFollowDomainName(profileSuggestion.domain)
         }
     }
 }
